Hide golang.org/x/text/language in exports of package alphabet.
In preparations to release a stable versions, we decide the public API must not export any unstable versioned APIs.
golang.org/x/text/language is currently in version 0.3.0. We can not guarantee it's API changes in the feature.
We want to be sure we can upgrade golang.org/x/text/language without changing our public API.

Package alphabet is not considered stable at the moment.
With version 1.0.0 we want to add it to our public API.